Completely refactor parser
Complete rewrite of parser.

Defines a cleaner datatype to represent the AST, based on the definition of a Basic Algebraic Expression (BAE) in the Joel Cohen Computer Algebra textbook.
Product is separate from Quotient, Power is now its own expression (not some kind of weird multi-argument function, as it was before).

Parses Sum and Product as n-ary operators, when they are not explicitly parenthesized.